+++ internal/dag/dag.go
@@ -1,211 +1,358 @@
 package dag
 
 import (
+	"context"
+	"database/sql"
+	"errors"
 	"fmt"
-	"sort"
+	"io/fs"
 	"strings"
-	"sync"
+
+	_ "github.com/mattn/go-sqlite3"
 )
 
-func New() *Graph {
-	return &Graph{
-		nodes: map[string]struct{}{},
-		ins:   map[string]map[string]struct{}{},
-		outs:  map[string]map[string]struct{}{},
+var ErrNotFound = fmt.Errorf("dcache: file not found")
+
+type File struct {
+	Path  string
+	Data  []byte
+	Mode  fs.FileMode
+	Links []string
+}
+
+type Link struct {
+	From string
+	To   string
+}
+
+type Cache interface {
+	Get(ctx context.Context, path string) (*File, error)
+	Put(ctx context.Context, path string, file *File) error
+	Delete(ctx context.Context, paths ...string) error
+	Close() error
+}
+
+func Load(ctx context.Context, dbPath string) (*cache, error) {
+	// Open the SQLite database
+	db, err := sql.Open("sqlite3", dbPath)
+	if err != nil {
+		return nil, err
 	}
-}
 
-type Graph struct {
-	mu    sync.RWMutex
-	nodes map[string]struct{}
-	ins   map[string]map[string]struct{}
-	outs  map[string]map[string]struct{}
-}
-
-func (g *Graph) Nodes() (nodes []string) {
-	for path := range g.nodes {
-		nodes = append(nodes, path)
+	// Create the files and links tables.
+	if _, err := db.ExecContext(ctx, `
+		CREATE TABLE IF NOT EXISTS files (
+			path TEXT PRIMARY KEY,
+			data BLOB,
+			mode INTEGER
+		)
+	`); err != nil {
+		return nil, err
 	}
-	sort.Strings(nodes)
-	return nodes
-}
-
-func (g *Graph) Set(path string) {
-	g.mu.Lock()
-	defer g.mu.Unlock()
-	g.nodes[path] = struct{}{}
-}
-
-func (g *Graph) Link(from, to string) {
-	// if from == to {
-	// 	return
-	// }
-	g.mu.Lock()
-	defer g.mu.Unlock()
-	// Create nodes if we haven't yet
-	g.nodes[from] = struct{}{}
-	g.nodes[to] = struct{}{}
-	// Set the dependencies
-	if g.outs[from] == nil {
-		g.outs[from] = map[string]struct{}{}
+	if _, err := db.ExecContext(ctx, `
+		CREATE TABLE IF NOT EXISTS links (
+			from_path TEXT,
+			to_path TEXT,
+			FOREIGN KEY(from_path) REFERENCES files(path),
+			FOREIGN KEY(to_path) REFERENCES files(path),
+			PRIMARY KEY(from_path, to_path)
+		)
+	`); err != nil {
+		return nil, err
 	}
-	g.outs[from][to] = struct{}{}
-	// Link the other way too
-	if g.ins[to] == nil {
-		g.ins[to] = map[string]struct{}{}
+
+	return &cache{db}, nil
+}
+
+type cache struct {
+	db *sql.DB
+}
+
+var _ Cache = (*cache)(nil)
+
+func (c *cache) Put(ctx context.Context, path string, file *File) error {
+	file.Path = path
+
+	// Start a transaction.
+	tx, err := c.db.Begin()
+	if err != nil {
+		return err
 	}
-	g.ins[to][from] = struct{}{}
-}
+	defer tx.Rollback()
 
-// Remove a node
-func (g *Graph) Remove(paths ...string) {
-	g.mu.Lock()
-	defer g.mu.Unlock()
-	for _, path := range paths {
-		g.remove(path)
+	// Insert the file into the files table.
+	if _, err := tx.Exec(`
+		INSERT INTO files (path, data, mode)
+		VALUES (?, ?, ?)
+	`, path, file.Data, file.Mode); err != nil {
+		return err
 	}
-}
 
-// Unlink and remove the node
-func (g *Graph) remove(path string) {
-	// Remove dependant links
-	for from := range g.ins[path] {
-		delete(g.ins[path], from)
-		delete(g.outs[from], path)
-	}
-	// Remove dependency links
-	for to := range g.outs[path] {
-		delete(g.outs[path], to)
-		delete(g.ins[to], path)
-	}
-	// Remove node
-	delete(g.nodes, path)
-}
-
-// Return the links out (dependencies)
-func (g *Graph) Children(from string) (tos []string) {
-	g.mu.RLock()
-	defer g.mu.RUnlock()
-	return g.children(from)
-}
-
-func (g *Graph) children(from string) (froms []string) {
-	for from := range g.outs[from] {
-		froms = append(froms, from)
-	}
-	sort.Strings(froms)
-	return froms
-}
-
-// Descendants recursively returns children, children of children, etc.
-// Descendants includes self.
-func (g *Graph) Descendants(path string) (descendants []string) {
-	g.mu.RLock()
-	defer g.mu.RUnlock()
-	descendants = append(descendants, path)
-	descendants = append(descendants, g.children(path)...)
-	return descendants
-}
-
-// Return the links in (parents)
-func (g *Graph) Parents(to string) (froms []string) {
-	g.mu.RLock()
-	defer g.mu.RUnlock()
-	return g.parents(to)
-}
-
-func (g *Graph) parents(to string) (tos []string) {
-	for to := range g.ins[to] {
-		tos = append(tos, to)
-	}
-	sort.Strings(tos)
-	return tos
-}
-
-// Ancestors recursively returns parents, parents of parents, etc.
-func (g *Graph) Ancestors(path string) (ancestors []string) {
-	ancestors = append(ancestors, g.parents(path)...)
-	return ancestors
-}
-
-// String returns a digraph
-func (g *Graph) String() string {
-	g.mu.RLock()
-	defer g.mu.RUnlock()
-	s := strings.Builder{}
-	s.WriteString("digraph g {\n")
-	// Rank from the bottom to the top
-	s.WriteString("  rankdir = BT\n")
-	for from := range g.nodes {
-		tos := g.outs[from]
-		if len(tos) == 0 {
-			s.WriteString(fmt.Sprintf("  %q\n", from))
-			continue
-		}
-		for to := range tos {
-			s.WriteString(fmt.Sprintf("  %q -> %q\n", from, to))
+	// Insert the links of the file into the links table.
+	for _, to := range file.Links {
+		if _, err := tx.Exec(`
+			INSERT INTO links (from_path, to_path)
+			VALUES (?, ?)
+		`, path, to); err != nil {
+			return err
 		}
 	}
-	s.WriteString("}\n")
-	return s.String()
+
+	return tx.Commit()
 }
 
-func (g *Graph) ShortestPath(start, end string) (nodes []string, err error) {
-	return g.shortestPath(start, end, nil)
-}
-
-func (g *Graph) shortestPath(start, end string, paths []string) (shortest []string, err error) {
-	if _, ok := g.nodes[start]; !ok {
-		return nil, fmt.Errorf("dag: %q doesn't exist", start)
-	}
-	paths = append(paths, start)
-	if start == end {
-		return paths, nil
-	}
-	for _, node := range g.children(start) {
-		if hasNode(paths, node) {
-			continue
+func (c *cache) Get(ctx context.Context, path string) (*File, error) {
+	row := c.db.QueryRow(`
+		SELECT path, data, mode
+		FROM files
+		WHERE path = ?
+	`, path)
+	var file File
+	if err := row.Scan(&file.Path, &file.Data, &file.Mode); err != nil {
+		if errors.Is(err, sql.ErrNoRows) {
+			return nil, fmt.Errorf("%w %q", ErrNotFound, path)
 		}
-		newPath, err := g.shortestPath(node, end, paths)
-		if err != nil {
+		return nil, err
+	}
+
+	// Retrieve the links of the file.
+	rows, err := c.db.Query(`
+		SELECT to_path
+		FROM links
+		WHERE from_path = ?
+	`, path)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+	for rows.Next() {
+		var to string
+		if err := rows.Scan(&to); err != nil {
 			return nil, err
 		}
-		lenNewPath := len(newPath)
-		lenShortest := len(shortest)
-		if lenNewPath == 0 {
-			continue
-		}
-		if lenShortest == 0 || lenNewPath < lenShortest {
-			shortest = newPath
-		}
+		file.Links = append(file.Links, to)
 	}
-	if len(shortest) == 0 {
-		return nil, fmt.Errorf("dag: no path between %q and %q", start, end)
+	if err := rows.Err(); err != nil {
+		return nil, err
 	}
-	return shortest, nil
+
+	return &file, nil
 }
 
-func hasNode(list []string, item string) bool {
-	for _, v := range list {
-		if v == item {
-			return true
-		}
+// Ancestors returns all the ancestor paths. Those that depend on the given
+// paths.
+func (c *cache) Ancestors(ctx context.Context, paths ...string) (deps []string, err error) {
+	// Start a transaction.
+	tx, err := c.db.Begin()
+	if err != nil {
+		return nil, err
 	}
-	return false
+	defer tx.Rollback()
+	deps, err = c.ancestors(ctx, tx, paths...)
+	if err != nil {
+		return nil, err
+	}
+	if err := tx.Commit(); err != nil {
+		return nil, err
+	}
+	return deps, nil
 }
 
-func (g *Graph) ShortestPathOf(start string, ends []string) (nodes []string, err error) {
-	for _, end := range ends {
-		shortest, err := g.shortestPath(start, end, nil)
-		if err != nil {
-			continue
-		}
-		if len(nodes) == 0 || len(shortest) < len(nodes) {
-			nodes = shortest
-		}
+func (c *cache) ancestors(ctx context.Context, tx *sql.Tx, paths ...string) (deps []string, err error) {
+	params := make([]interface{}, len(paths))
+	for i, path := range paths {
+		params[i] = path
 	}
-	if len(nodes) == 0 {
-		return nil, fmt.Errorf("dag: no path between %q and %v", start, ends)
+	// Use a recursive CTE to retrieve the deps of the file.
+	sql := `
+		WITH RECURSIVE deps AS (
+			SELECT from_path
+			FROM links
+			WHERE to_path IN ` + parameterize(len(params)) + `
+			UNION ALL
+			SELECT links.from_path
+			FROM links
+			JOIN deps
+			ON links.to_path = deps.from_path
+		)
+		SELECT DISTINCT from_path
+		FROM deps
+		ORDER BY from_path
+	`
+	rows, err := tx.QueryContext(ctx, sql, params...)
+
+	if err != nil {
+		return nil, err
 	}
-	return nodes, nil
+	defer rows.Close()
+	for rows.Next() {
+		var to string
+		if err := rows.Scan(&to); err != nil {
+			return nil, err
+		}
+		deps = append(deps, to)
+	}
+	if err := rows.Err(); err != nil {
+		return nil, err
+	}
+
+	return deps, nil
+}
+
+func (c *cache) Delete(ctx context.Context, paths ...string) error {
+	// Start a transaction
+	tx, err := c.db.Begin()
+	if err != nil {
+		return err
+	}
+	defer tx.Rollback()
+
+	// Get all the ancestors paths
+	ancestors, err := c.ancestors(ctx, tx, paths...)
+	if err != nil {
+		return err
+	}
+	params := make([]interface{}, len(paths)+len(ancestors))
+	for i, path := range paths {
+		params[i] = path
+	}
+	for i, path := range ancestors {
+		params[i+len(paths)] = path
+	}
+
+	// Delete all paths and their dependencies
+	sql := `
+		DELETE FROM files
+		WHERE path IN ` + parameterize(len(params)) + `
+	`
+	_, err = tx.ExecContext(ctx, sql, params...)
+	if err != nil {
+		return err
+	}
+
+	// Delete all links to these paths
+	sql = `
+		DELETE FROM links
+		WHERE to_path IN ` + parameterize(len(params)) + `
+		OR from_path IN ` + parameterize(len(params)) + `
+	`
+	_, err = tx.ExecContext(ctx, sql, append(params, params...)...)
+	if err != nil {
+		return err
+	}
+
+	return tx.Commit()
+}
+
+func parameterize(n int) string {
+	out := "("
+	for i := 0; i < n; i++ {
+		if i > 0 {
+			out += ", "
+		}
+		out += "?"
+	}
+	out += ")"
+	return out
+}
+
+func (c *cache) Files(ctx context.Context) ([]*File, error) {
+	rows, err := c.db.QueryContext(ctx, `
+		SELECT path, data, mode
+		FROM files
+		ORDER BY path
+	`)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+	var files []*File
+	for rows.Next() {
+		var file File
+		if err := rows.Scan(&file.Path, &file.Data, &file.Mode); err != nil {
+			return nil, err
+		}
+		files = append(files, &file)
+	}
+	if err := rows.Err(); err != nil {
+		return nil, err
+	}
+	return files, nil
+}
+
+func (c *cache) Links(ctx context.Context) ([]*Link, error) {
+	rows, err := c.db.QueryContext(ctx, `
+		SELECT from_path, to_path
+		FROM links
+		ORDER BY from_path, to_path
+	`)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+	var links []*Link
+	for rows.Next() {
+		var link Link
+		if err := rows.Scan(&link.From, &link.To); err != nil {
+			return nil, err
+		}
+		links = append(links, &link)
+	}
+	if err := rows.Err(); err != nil {
+		return nil, err
+	}
+	return links, nil
+}
+
+// Print the DAG in a dot graph format. That you can paste in here:
+// https://dreampuf.github.io/GraphvizOnline
+func (c *cache) Print(ctx context.Context) (string, error) {
+	out := new(strings.Builder)
+	// Open the digraph
+	out.WriteString("digraph dag {\n")
+	// Print the nodes of the graph.
+	rows, err := c.db.Query(`
+		SELECT path
+		FROM files
+	`)
+	if err != nil {
+		return "", err
+	}
+	defer rows.Close()
+	for rows.Next() {
+		var path string
+		if err := rows.Scan(&path); err != nil {
+			return "", err
+		}
+		fmt.Fprintf(out, "\t%[1]q\n", path)
+	}
+	if err := rows.Err(); err != nil {
+		return "", err
+	}
+	// Print the edges of the graph.
+	rows, err = c.db.Query(`
+		SELECT from_path, to_path
+		FROM links
+	`)
+	if err != nil {
+		return "", err
+	}
+	defer rows.Close()
+	for rows.Next() {
+		var from, to string
+		if err := rows.Scan(&from, &to); err != nil {
+			return "", err
+		}
+		fmt.Fprintf(out, "\t%q -> %q\n", from, to)
+	}
+	if err := rows.Err(); err != nil {
+		return "", err
+	}
+	// Close the graph.
+	out.WriteString("}\n")
+	return out.String(), nil
+}
+
+func (c *cache) Close() error {
+	return c.db.Close()
 }
